Refactor application to modular package structure
Reorganize monolithic main.py (278 lines) into clean, maintainable
package architecture following separation of concerns principles.

## Changes

### Structure
- Create weather_app/ package with organized modules
- Separate concerns: web, storage, db, config layers
- Reduce main.py from 278 lines to 11 lines

### Modules Created
- weather_app/web/ - FastAPI app factory, routes, models
- weather_app/storage/ - Repository pattern for data access
- weather_app/db/ - Database connection management
- weather_app/config.py - Enhanced configuration module
- weather_app/fetch/ - Placeholder for future CLI features
- weather_app/utils/ - Placeholder for utilities

### Improvements
- Fix FastAPI deprecation warning (regex → pattern parameter)
- Improve error handling (ValueError vs RuntimeError vs HTTPException)
- Add type hints throughout
- Better path handling using pathlib
- Environment-driven HOST/PORT configuration
- Add future-ready retention policy configs

### Backward Compatibility
- All API endpoints unchanged (/, /weather, /weather/latest, /weather/stats)
- Identical response format
- Same environment variables (USE_TEST_DB, etc.)
- Database schema unchanged

### Testing
- Verified import tests pass
- Side-by-side comparison shows identical behavior
- All endpoints tested and working

### Documentation
- ARCHITECTURE.md - Design diagrams and data flow
- BEFORE_AFTER_COMPARISON.md - Side-by-side code comparison
- REFACTORING_SUMMARY.md - Complete refactoring details
- REFACTORING_CHECKLIST.md - Testing and review checklist
- REVIEW_ME_FIRST.md - Quick reference guide

### Backup Files
- main_old.py - Original monolithic implementation
- config_old.py - Original configuration
