+++ REFACTORING_CHECKLIST.md
@@ -1,0 +1,210 @@
+# Refactoring Review Checklist
+
+## What Was Created
+
+### New Package Structure ✓
+```
+weather_app/
+├── __init__.py                    - Package initialization
+├── config.py                      - Configuration (refactored from root)
+├── db/
+│   ├── __init__.py
+│   └── session.py                 - Database connection management
+├── web/
+│   ├── __init__.py
+│   ├── app.py                     - FastAPI app factory
+│   ├── models.py                  - Pydantic response models
+│   └── routes.py                  - API endpoints (extracted from main.py)
+├── storage/
+│   ├── __init__.py
+│   └── repository.py              - Database query abstraction
+├── fetch/
+│   └── __init__.py                - Placeholder for future fetcher logic
+└── utils/
+    └── __init__.py                - Placeholder for future utilities
+```
+
+### New Files Created
+
+| File | Purpose | Status |
+|------|---------|--------|
+| `weather_app/__init__.py` | Package marker | ✓ Created |
+| `weather_app/config.py` | Configuration (moved & enhanced) | ✓ Created |
+| `weather_app/db/session.py` | DB connection management | ✓ Created |
+| `weather_app/web/app.py` | FastAPI app factory | ✓ Created |
+| `weather_app/web/models.py` | Pydantic models | ✓ Created |
+| `weather_app/web/routes.py` | API endpoints | ✓ Created |
+| `weather_app/storage/repository.py` | Data access layer | ✓ Created |
+| `main_refactored.py` | New entry point | ✓ Created |
+| `REFACTORING_SUMMARY.md` | This refactoring summary | ✓ Created |
+| `ARCHITECTURE.md` | Architecture diagrams | ✓ Created |
+
+### What Was Extracted from main.py
+
+- ✓ `WeatherData` model → `weather_app/web/models.py`
+- ✓ `DatabaseStats` model → `weather_app/web/models.py`
+- ✓ Database helper functions → `weather_app/db/session.py`
+- ✓ All 4 API endpoints → `weather_app/web/routes.py`
+- ✓ Pydantic model definitions → `weather_app/web/models.py`
+
+### Improvements Made
+
+- ✓ **Fixed deprecation warning** - Changed `regex=` to `pattern=` in FastAPI
+- ✓ **Better error handling** - Repository layer validates and converts errors
+- ✓ **Cleaner imports** - Each module has clear dependencies
+- ✓ **Config enhancements** - Added future-ready config options
+- ✓ **Application factory pattern** - `create_app()` for better testing
+- ✓ **Separated concerns** - Routes don't contain query logic
+
+---
+
+## Review Questions
+
+### 1. Code Organization
+- [ ] Is the directory structure clear and logical?
+- [ ] Does each module have a single responsibility?
+- [ ] Are dependencies between modules minimal and clear?
+
+### 2. Functionality
+- [ ] Does `main_refactored.py` serve the same purpose as `main.py`?
+- [ ] Are all 4 API endpoints preserved?
+- [ ] Is the database behavior identical?
+
+### 3. Configuration
+- [ ] Are the new config options (`FULL_RESOLUTION_YEARS`, etc.) acceptable?
+- [ ] Should default values be different?
+- [ ] Is the config precedence correct (env vars → .env → defaults)?
+
+### 4. Future-Readiness
+- [ ] Is the structure ready for CLI commands?
+- [ ] Can fetch logic be added to `weather_app/fetch/` without conflicts?
+- [ ] Will aggregation/retention fit cleanly in `weather_app/storage/`?
+
+### 5. Migration Strategy
+- [ ] Keep both `main.py` and `main_refactored.py` during testing? [Y/N]
+- [ ] Should we add imports in `main.py` to delegate to new structure?
+- [ ] When ready to commit, rename `main_refactored.py` → `main.py`?
+
+---
+
+## Testing the Refactored Code
+
+### Option A: Quick Validation
+```powershell
+# Test that imports work
+python -c "from weather_app.web.app import create_app; app = create_app(); print(f'Routes: {len(app.routes)}')"
+```
+
+### Option B: Full Server Test
+```powershell
+# Set test database
+$env:USE_TEST_DB="true"
+
+# First, generate test data
+python generate_test_data.py --days 7
+
+# Run the refactored server
+python main_refactored.py
+
+# In another terminal:
+curl http://localhost:8000/weather/latest
+```
+
+### Option C: Compare Both Versions
+```powershell
+# Terminal 1: Old version
+$env:USE_TEST_DB="true"
+python main.py
+
+# Terminal 2: New version (different port)
+$env:USE_TEST_DB="true"
+$env:BIND_PORT="8001"
+python main_refactored.py
+
+# Then test both APIs:
+curl http://localhost:8000/weather/stats
+curl http://localhost:8001/weather/stats
+# Should return identical results
+```
+
+---
+
+## Next Steps (Pending Approval)
+
+### Phase 1: Review & Approval
+- [ ] Review structure and ask questions
+- [ ] Suggest any changes to organization
+- [ ] Approve before proceeding
+
+### Phase 2: Testing (After Approval)
+- [ ] Run `main_refactored.py` with test database
+- [ ] Test all 4 API endpoints
+- [ ] Verify identical behavior to `main.py`
+- [ ] Test with production database
+
+### Phase 3: Migration (After Testing)
+- [ ] Backup old files:
+  - `main.py` → `main_old.py`
+  - `config.py` → `config_old.py`
+- [ ] Rename `main_refactored.py` → `main.py`
+- [ ] Update documentation to reference new structure
+- [ ] Delete old files after confirming everything works
+
+### Phase 4: Enhancement (After Migration)
+1. **Add CLI Module** - Convert scripts to Click commands
+   ```python
+   weather-app fetch
+   weather-app backfill
+   weather-app aggregate
+   weather-app export
+   ```
+
+2. **Add Fetch Module** - Move API client and fetcher logic
+   - `weather_app/fetch/ambient_client.py`
+   - `weather_app/fetch/fetcher.py`
+
+3. **Add Retention Logic** - Implement aggregation and cleanup
+   - `weather_app/storage/retention.py`
+   - `weather_app/storage/aggregator.py`
+
+4. **Add Tests** - Build pytest suite
+   - `tests/unit/`
+   - `tests/integration/`
+
+---
+
+## Files Reference
+
+### Documentation Created
+- `REFACTORING_SUMMARY.md` - Overview of changes
+- `ARCHITECTURE.md` - Detailed architecture diagrams
+
+### Old Code (Preserved)
+- `main.py` - Original (keep during testing)
+- `config.py` - Original (keep during testing)
+
+### New Code (Ready for Testing)
+- `main_refactored.py` - New entry point
+- `weather_app/` - All new modules
+
+---
+
+## Questions or Concerns?
+
+1. **Module organization** - Is there a better way to structure the modules?
+2. **Dependencies** - Should we use SQLAlchemy now or keep raw SQLite?
+3. **Config** - Any config options that should be added/removed?
+4. **Testing** - Should we add pytest now or wait?
+5. **CLI** - Should we implement Click CLI in Phase 2 or Phase 4?
+
+---
+
+## Verification
+
+Import test passed ✓
+- All modules import successfully
+- FastAPI app creates with 8 routes
+- Config loads correctly
+- Database detection works (PRODUCTION/TEST mode)
+
+Ready for your review!
