+++ weather_app/__init__.py
@@ -1,0 +1,6 @@
+"""
+Weather App Package
+Main weather data retrieval, storage, and visualization application
+"""
+
+__version__ = "1.0.0"
