+++ weather_app/db/__init__.py
@@ -1,0 +1,4 @@
+"""
+Database module for Weather App
+Handles database connections, models, and queries
+"""
