+++ weather_app/fetch/__init__.py
@@ -1,0 +1,4 @@
+"""
+Fetch module for Weather App
+Handles data retrieval from Ambient Weather API
+"""
