+++ weather_app/storage/__init__.py
@@ -1,0 +1,4 @@
+"""
+Storage module for Weather App
+Handles database operations and data persistence
+"""
