+++ weather_app/utils/__init__.py
@@ -1,0 +1,4 @@
+"""
+Utilities module for Weather App
+Contains logging, retry logic, and helper functions
+"""
