+++ weather_app/web/__init__.py
@@ -1,0 +1,4 @@
+"""
+Web module for Weather App
+Contains FastAPI app, routes, and models
+"""
